requirements.txt
fastapi
uvicorn
pydantic
groq
pinecone-client
supabase
python-dotenv
langchain
langchain-community
python-multipart
requests
